using timeline semaphores and creating queues types with vulkan
Still need to correctly find a way to use timeline semaphores correctly. They emulate the DirectX12 fences, so it's a perfect fit if it works.

In addition, removing semaphore types. As vulkan will now only use timeline semaphores instead of normal and fences. The exception is for the Window integration API to get and present images.